fix: wip, visualizing the map markers
