+++ src/app/map/MapComponent.tsx
@@ -8,7 +8,8 @@ import { Marker } from "react-map-gl";
 import axios from "axios";
 import Image from "next/legacy/image";
 import LoadingComponent from "@components/LoadingComponent";
-import Head from "next/head";
+import { getValidPosts, sortPosts } from "app/feed/FeedComponent";
+
 const MapScreen: React.FC = () => {
   const [audioNFTs, setAudioNFTs] = useState<AudioNFT[] | undefined>(undefined);
   const [showModal, setShowModal] = useState<boolean>(false);
@@ -56,43 +57,57 @@ const MapScreen: React.FC = () => {
   };
 
   const fetchNFTs = async () => {
+    const response = await fetch("/api/nfts?initialPageNumber=1");
     try {
-      const response = await fetch("/api/nfts?initialPageNumber=1");
-      const result = await response.json();
-
-      if (response.ok) {
-        const postPromises: Promise<AudioNFT>[] = result.map(
-          async (item: any) => {
-            const { animationUrl, attributes } = await fetchJsonData(
-              item.content.json_uri
-            );
-            const owner = item.ownership.owner;
-            const metadata = item.content.metadata;
-            const assetId = item.id;
-            const burnt = item.burnt;
-
-            const attributesObj = attributes?.reduce(
-              (acc: any, attribute: any) => {
-                acc[attribute.trait_type] = attribute.value;
-                return acc;
-              },
-              {}
-            );
-            if (animationUrl && attributes && !burnt) {
-              return { animationUrl, attributesObj, metadata, owner, assetId };
-            }
-          }
-        );
-        const validPosts = (await Promise.all(postPromises)).filter(
-          (post) => post !== undefined
-        );
-        setAudioNFTs(validPosts);
+      const validPosts = await getValidPosts(response);
+      if (validPosts) {
+        const posts = sortPosts(validPosts);
+        setAudioNFTs(posts);
         checkIfAudioNFTisShared(validPosts);
       }
     } catch (error) {
       console.error("Error fetching assets:", error);
     }
   };
+
+  // const fetchNFTs = async () => {
+  //   try {
+  //     const response = await fetch("/api/nfts?initialPageNumber=1");
+  //     const result = await response.json();
+
+  //     if (response.ok) {
+  //       const postPromises: Promise<AudioNFT>[] = result.map(
+  //         async (item: any) => {
+  //           const { animationUrl, attributes } = await fetchJsonData(
+  //             item.content.json_uri
+  //           );
+  //           const owner = item.ownership.owner;
+  //           const metadata = item.content.metadata;
+  //           const assetId = item.id;
+  //           const burnt = item.burnt;
+
+  //           const attributesObj = attributes?.reduce(
+  //             (acc: any, attribute: any) => {
+  //               acc[attribute.trait_type] = attribute.value;
+  //               return acc;
+  //             },
+  //             {}
+  //           );
+  //           if (animationUrl && attributes && !burnt) {
+  //             return { animationUrl, attributesObj, metadata, owner, assetId };
+  //           }
+  //         }
+  //       );
+  //       const validPosts = (await Promise.all(postPromises)).filter(
+  //         (post) => post !== undefined
+  //       );
+  //       setAudioNFTs(validPosts);
+  //       checkIfAudioNFTisShared(validPosts);
+  //     }
+  //   } catch (error) {
+  //     console.error("Error fetching assets:", error);
+  //   }
+  // };
 
   const showSharedNFT = useCallback(
     async (audioNFTid: string, audioNFTs: AudioNFT[]) => {
@@ -141,6 +156,7 @@ const MapScreen: React.FC = () => {
     if (!audioNFTs || audioNFT?.status === "pending") {
       fetchNFTs();
     }
+    console.log("AudioNFTs: ", audioNFTs);
   }, [checkIfAudioNFTisShared, audioNFTs, audioNFT?.status]);
 
   const markers: JSX.Element[] = useMemo(
@@ -149,15 +165,15 @@ const MapScreen: React.FC = () => {
         ? audioNFTs
             .filter(
               (audioNFT) =>
-                audioNFT.attributesObj &&
-                audioNFT.attributesObj.Long !== undefined &&
-                audioNFT.attributesObj.Lat !== undefined
+                audioNFT.props &&
+                audioNFT.props.latitude !== undefined &&
+                audioNFT.props.longitude !== undefined
             )
             .map((audioNFT, index) => (
               <Marker
                 key={index}
-                longitude={Number(audioNFT.attributesObj.Long)}
-                latitude={Number(audioNFT.attributesObj.Lat)}
+                longitude={Number(audioNFT.props.longitude)}
+                latitude={Number(audioNFT.pros.latitude)}
                 color="red"
                 onClick={() => {
                   setAudioNFT(audioNFT);
@@ -178,33 +194,6 @@ const MapScreen: React.FC = () => {
 
   return (
     <>
-      <Head>
-        <title>øhr</title>
-        <meta name="twitter:card" content="summary_large_image" />
-        <meta name="twitter:site" content="@ohr_xyz" />
-        <meta
-          name="twitter:title"
-          content={audioNFT ? audioNFT.attributesObj.Vibe : "øhr"}
-        />
-        <meta name="twitter:description" content="listen, mint, share" />
-        <meta
-          property="og:url"
-          content={
-            audioNFT
-              ? `https://ohr-app.xyz/map?id=${audioNFT.id}&latitude=${audioNFT.attributesObj.Lat}&longitude=${audioNFT.attributesObj.Long}`
-              : "https://ohr-app.xyz/"
-          }
-        />
-        <meta
-          name="twitter:image"
-          content="https://raw.githubusercontent.com/noamrubin22/ohr_xyz/main/src/assets/link-preview-img.png"
-        />
-        <meta property="og:description" content="listen, mint, share" />
-        <meta
-          property="og:image"
-          content="https://raw.githubusercontent.com/noamrubin22/ohr_xyz/main/src/assets/link-preview-img.png"
-        />
-      </Head>
       <LayoutComponent showTitle="Explore" showFooter={true} showNavBar={true}>
         <h2 className="text-center text-sm">
           {" "}
